Fix vector and complex loop tests for x86

+++ solidity/contract/loop/complex_2.sol
@@ -19,7 +19,8 @@ pragma solidity ^0.8.0;
 
 contract Test {
     function entry(uint8 param) private pure returns(uint64) {
-        uint8 cnt = 0;
+        uint64 cnt = 0;
+
         uint8 j = 1;
         while (j < 4) {
             uint8 p = 0;
@@ -49,6 +50,7 @@ contract Test {
             } while (true);
             j *= 2;
         }
+        
         return cnt;
     }
 }
